Revert "fix(module-tools): change the output css module json filename (#4774)"
This reverts commit bd7b116b5c9ab306262e0452e61fddc97fba5b70.
